split Dataframe class into modules
