+++ .rubocop_todo.yml
@@ -1,10 +1,18 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2022-08-16 13:20:50 UTC using RuboCop version 1.35.0.
+# on 2024-03-03 13:59:21 UTC using RuboCop version 1.60.2.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
+
+# Offense count: 1
+# This cop supports safe autocorrection (--autocorrect).
+# Configuration parameters: EnforcedStyle, IndentationWidth.
+# SupportedStyles: aligned, indented
+Layout/LineEndStringConcatenationIndentation:
+  Exclude:
+    - 'lib/daru_lite/data_frame/indexable.rb'
 
 # Offense count: 1
 # Configuration parameters: AllowComments.
@@ -13,6 +21,7 @@ Lint/EmptyClass:
     - 'lib/daru_lite/accessors/mdarray_wrapper.rb'
 
 # Offense count: 5
+# Configuration parameters: AllowedParentClasses.
 Lint/MissingSuper:
   Exclude:
     - 'lib/daru_lite/date_time/offsets.rb'
@@ -20,61 +29,50 @@ Lint/MissingSuper:
     - 'lib/daru_lite/index/index.rb'
     - 'lib/daru_lite/index/multi_index.rb'
 
-# Offense count: 6
+# Offense count: 5
+# This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: CheckForMethodsWithNoSideEffects.
 Lint/Void:
   Exclude:
     - 'lib/daru_lite/category.rb'
-    - 'lib/daru_lite/dataframe.rb'
+    - 'lib/daru_lite/data_frame/indexable.rb'
     - 'lib/daru_lite/vector.rb'
 
-# Offense count: 40
-# Configuration parameters: AllowedMethods, AllowedPatterns, IgnoredMethods, CountRepeatedAttributes.
+# Offense count: 41
+# Configuration parameters: AllowedMethods, AllowedPatterns, CountRepeatedAttributes.
 Metrics/AbcSize:
   Max: 34
 
-# Offense count: 3
+# Offense count: 5
 # Configuration parameters: CountComments, CountAsOne.
 Metrics/ClassLength:
-  Max: 189
+  Max: 188
 
 # Offense count: 6
-# Configuration parameters: AllowedMethods, AllowedPatterns, IgnoredMethods.
+# Configuration parameters: AllowedMethods, AllowedPatterns.
 Metrics/CyclomaticComplexity:
   Max: 9
 
-# Offense count: 61
-# Configuration parameters: CountComments, CountAsOne, ExcludedMethods, AllowedMethods, AllowedPatterns, IgnoredMethods.
+# Offense count: 60
+# Configuration parameters: CountComments, CountAsOne, AllowedMethods, AllowedPatterns.
 Metrics/MethodLength:
   Max: 15
 
-# Offense count: 2
+# Offense count: 4
 # Configuration parameters: CountComments, CountAsOne.
 Metrics/ModuleLength:
   Max: 190
 
 # Offense count: 4
-# Configuration parameters: AllowedMethods, AllowedPatterns, IgnoredMethods.
+# Configuration parameters: AllowedMethods, AllowedPatterns.
 Metrics/PerceivedComplexity:
   Max: 10
 
-# Offense count: 72
+# Offense count: 66
 # Configuration parameters: MinNameLength, AllowNamesEndingInNumbers, AllowedNames, ForbiddenNames.
-# AllowedNames: at, by, db, id, in, io, ip, of, on, os, pp, to
+# AllowedNames: as, at, by, cc, db, id, if, in, io, ip, of, on, os, pp, to
 Naming/MethodParameterName:
-  Exclude:
-    - 'lib/daru_lite/category.rb'
-    - 'lib/daru_lite/core/group_by.rb'
-    - 'lib/daru_lite/core/merge.rb'
-    - 'lib/daru_lite/core/query.rb'
-    - 'lib/daru_lite/dataframe.rb'
-    - 'lib/daru_lite/date_time/index.rb'
-    - 'lib/daru_lite/date_time/offsets.rb'
-    - 'lib/daru_lite/extensions/which_dsl.rb'
-    - 'lib/daru_lite/io/io.rb'
-    - 'lib/daru_lite/maths/statistics/dataframe.rb'
-    - 'lib/daru_lite/maths/statistics/vector.rb'
-    - 'lib/daru_lite/vector.rb'
+  Enabled: false
 
 # Offense count: 5
 # Configuration parameters: NamePrefix, ForbiddenPrefixes, AllowedMethods, MethodDefinitionMacros.
@@ -85,13 +83,14 @@ Naming/MethodParameterName:
 Naming/PredicateName:
   Exclude:
     - 'spec/**/*'
-    - 'lib/daru_lite/dataframe.rb'
+    - 'lib/daru_lite/data_frame/missable.rb'
+    - 'lib/daru_lite/data_frame/queryable.rb'
     - 'lib/daru_lite/vector.rb'
 
 # Offense count: 5
 Security/MarshalLoad:
   Exclude:
-    - 'lib/daru_lite/dataframe.rb'
+    - 'lib/daru_lite/data_frame/i_o_able.rb'
     - 'lib/daru_lite/date_time/index.rb'
     - 'lib/daru_lite/index/index.rb'
     - 'lib/daru_lite/io/io.rb'
@@ -102,7 +101,7 @@ Style/ClassVars:
   Exclude:
     - 'lib/daru_lite.rb'
 
-# Offense count: 44
+# Offense count: 58
 # Configuration parameters: AllowedConstants.
 Style/Documentation:
   Enabled: false
@@ -113,6 +112,10 @@ Style/MapToHash:
   Exclude:
     - 'lib/daru_lite/category.rb'
     - 'lib/daru_lite/core/group_by.rb'
+    - 'lib/daru_lite/data_frame/convertible.rb'
+    - 'lib/daru_lite/data_frame/duplicatable.rb'
+    - 'lib/daru_lite/data_frame/fetchable.rb'
+    - 'lib/daru_lite/data_frame/joinable.rb'
     - 'lib/daru_lite/dataframe.rb'
 
 # Offense count: 1
@@ -125,7 +128,7 @@ Style/MultilineBlockChain:
 # AllowedMethods: respond_to_missing?
 Style/OptionalBooleanParameter:
   Exclude:
-    - 'lib/daru_lite/dataframe.rb'
+    - 'lib/daru_lite/data_frame/convertible.rb'
     - 'lib/daru_lite/maths/statistics/vector.rb'
     - 'lib/daru_lite/vector.rb'
 
@@ -133,5 +136,4 @@ Style/OptionalBooleanParameter:
 # This cop supports unsafe autocorrection (--autocorrect-all).
 Style/RedundantSelfAssignment:
   Exclude:
-    - 'lib/daru_lite/dataframe.rb'
-
+    - 'lib/daru_lite/data_frame/joinable.rb'
